# Use a precompiled regex for strip_color (2.7 to 4x faster)