This speeds up `strip_color` and `ansilen` by precompiling the regex.

### String without color code :

* Before :

```
python -m timeit -s 'import re; s = "abcdefghijklmnopqrstuvwxyz"' 're.sub("\x1b\\[(K|.*?m)", "", s)'
500000 loops, best of 5: 822 nsec per loop
```

* After :

```
python -m timeit -s 'import re; s = "abcdefghijklmnopqrstuvwxyz"; r = re.compile("\x1b\\[(K|.*?m)")' 'r.sub("", s)'
1000000 loops, best of 5: 202 nsec per loop
```

**=> ~4x speedup**

### String with color code :

* Before :

```
python -m timeit -s 'import re; s = "abcdefghij\x1b[30mklmnop\1b[0mqrstuvwxyz"' 're.sub("\x1b\\[(K|.*?m)", "", s)'
200000 loops, best of 5: 1.47 usec per loop
```

* After :

```
python -m timeit -s 'import re; s = "abcdefghij\x1b[30mklmnop\1b[0mqrstuvwxyz"; r = re.compile("\x1b\\[(K|.*?m)")' 'r.sub("", s)'
500000 loops, best of 5: 547 nsec per loop
```

**=> ~2.7x speedup**